LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigation-traverseTo-top-cancels-cross-document-child-expected.txt



FAIL navigate.traverseTo() cancelled by top frame cancels cross-document iframe assert_equals: expected 2 but got 1
